pyproject.toml
[tool.poetry]
name = "hello-fast"
version = "0.1.0"
description = ""
authors = ["el hassane gargem <elhassanegargem@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.79.0"
psycopg2-binary = "2.9.1"
sqlmodel = "0.0.4"
SQLAlchemy-Utils = "^0.38.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
